- fixed bug - separated widgets into each folder - made sceens folder and shared folder
