+++ CONTRIBUTORS.md
@@ -1,1 +1,2 @@
 <!-- [name](github profile link) -->
+[Darshan Rander](https://github.com/SirusCodes/)
